pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidianite"
version = "0.1.0"
description = "Sync your Obsidian Vault with a private GitHub repository"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Obsidianite", email = "noreply@example.com" }
]
dependencies = [
  "typer[all]>=0.12.3",
  "GitPython>=3.1.43",
  "PyGithub>=2.3.0",
  "python-dotenv>=1.0.1",
  "colorama>=0.4.6",
  "rich>=13.7.1"
]

[project.urls]
Homepage = "https://github.com/"

[project.scripts]
obsidianite = "obsidianite.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]


